+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
 - Add support for getting license
 - Add support for Raw Authentication in VST (support external jwt token as raw element)
+- Fix race when using WithRawResponse/WithResponse context with agencyConnection 
 
 ## [1.6.0](https://github.com/arangodb/go-driver/tree/v1.6.0) (2023-05-30)
 - Add ErrArangoDatabaseNotFound and IsExternalStorageError helper to v2